- Reorganized utilities and a few other things to have an extra catchall file with functions such as clusterMatrix that are general but accessible to users - Move signdy related functions to signature.py
