+++ prody/utilities/misctools.py
@@ -1,12 +1,10 @@
 """This module defines miscellaneous utility functions."""
 
 from numpy import unique, linalg, diag, sqrt, dot
-import scipy.cluster.hierarchy as sch
-from scipy import spatial
 
 __all__ = ['Everything', 'rangeString', 'alnum', 'importLA', 'dictElement',
            'intorfloat', 'startswith', 'showFigure', 'countBytes', 'sqrtm',
-           'clusterMatrix', 'saxsWater', 'count']
+           'saxsWater', 'count']
 
 
 class Everything(object):
@@ -178,51 +176,6 @@ def getMasses(elements):
 
 def count(L, a=None):
     return len([b for b in L if b is a])
-
-def clusterMatrix(similarity_matrix=None, distance_matrix=None, labels=None, no_plot=True, **kwargs):
-    """
-    Cluster a similarity matrix or a distance matrix using scipy.cluster.hierarchy and 
-    return the indices, sorted matrix, sorted labels, and sorting dendrogram dict.
-    
-    :arg similarity_matrix: an N-by-N matrix containing some measure of similarity 
-        such as sequence identity, mode-mode overlap, or spectral overlap
-    :type similarity_matrix: array
-    
-    :arg distance_matrix: an N-by-N matrix containing some measure of distance 
-        such as 1. - seqid_matrix, rmsds, or distances in PCA space
-    :type similarity_matrix: array
-    
-    :arg labels: labels for each matrix row that can be returned sorted
-    :type labels: list
-    
-    Other arguments for scipy.hierarchy.linkage and scipy.hierarchy.dendrogram
-        can also be provided and will be taken as kwargs.
-        
-    :arg no_plot: If True, don't plot the dendrogram.
-        default is True
-    :type no_plot: bool
-    """
-    if similarity_matrix is None and distance_matrix is None:
-        raise ValueError('Please provide a similarity matrix or a distance matrix')
-    elif distance_matrix is None:
-        distance_matrix = 1. - similarity_matrix
-    
-    orientation = kwargs.pop('orientiation','right')
-    
-    formatted_distance_matrix = spatial.distance.squareform(distance_matrix)
-    linkage_matrix = sch.linkage(formatted_distance_matrix, **kwargs)
-    sorting_dendrogram = sch.dendrogram(linkage_matrix, orientation=orientation, labels=labels, no_plot=no_plot)
-
-    indices = sorting_dendrogram['leaves']
-    sorted_labels = sorting_dendrogram['ivl']
-    
-    if similarity_matrix is None:
-        sorted_matrix = distance_matrix[indices,:]
-    else:
-        sorted_matrix = similarity_matrix[indices,:]
-    sorted_matrix = sorted_matrix[:,indices]
-    
-    return indices, sorted_matrix, sorted_labels, sorting_dendrogram
 
 def saxsWater():
     return [[-48.347,-49.439,-50.456],
